+++ src/types/identity.ts
@@ -1,3 +1,5 @@
+import type { Store } from "pinia"
+
 export interface IdentityState {
     birthname?: string
     firstname?: string
@@ -10,3 +12,9 @@ export interface IdentityStore {
     getFullname(): string
     setData(identity: IdentityState): void
 }
+
+export interface IdentityGetters {
+    readonly fullname: string
+}
+
+export type IdentityStoreInstance = Store<string, Omit<IdentityState, 'fullname'>> & IdentityGetters & IdentityStore