+++ src/types/resourceId.ts
@@ -1,3 +1,5 @@
+import type { Store } from "pinia";
+
 export interface ResourceId {
   "@id"?: string;
   id?: number | string;
@@ -6,3 +8,5 @@ export interface ResourceId {
 export interface ResourceIdStore {
   setData: (data: ResourceId) => void
 }
+
+export type ResourceIdStoreInstance = Store<string, ResourceId> & ResourceIdStore