+++ src/types/webuser.ts
@@ -1,4 +1,5 @@
 import type { ResourceId, ResourceIdStore } from "./resourceId"
+import type { ExtendedStoreInstance } from "./store"
 
 export interface WebUserStore extends ResourceIdStore {
     setData: (data: Partial<WebUserState>) => void
@@ -9,3 +10,5 @@ export interface WebUserState extends ResourceId {
     password?: string
     username?: string
 }
+
+export type WebUserStoreInstance = ExtendedStoreInstance<WebUserState, WebUserStore>